requirements.txt
absl-py==2.2.2
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
astunparse==1.6.3
attrs==25.3.0
autograd==1.8.0
autopy==4.0.1
blinker==1.9.0
blis==1.2.0
bokeh==3.4.3
breadability==0.1.20
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
distro==1.9.0
dm-tree==0.1.8
docopt==0.6.2
et_xmlfile==2.0.0
exceptiongroup==1.3.0
executing==2.2.0
Farama-Notifications==0.0.4
fastapi==0.115.12
filelock==3.18.0
Flask==2.3.2
flask-cors==5.0.1
Flask-Login==0.6.3
flatbuffers==25.2.10
fonttools==4.58.0
fsspec==2025.3.2
gast==0.4.0
gensim==4.3.3
google-auth==2.40.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.71.0
gym==0.26.2
gym-notices==0.0.8
gymnasium==1.1.1
h11==0.16.0
h5py==3.13.0
HandTrackingModule==0.2
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.32.2
idna==3.10
imageio==2.37.0
imgaug==0.4.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.18.1
itsdangerous==2.2.0
jax==0.4.30
jaxlib==0.4.30
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.9.2
Keras-Preprocessing==1.1.2
kiwisolver==1.4.7
lazy_loader==0.4
libclang==18.1.1
lxml==5.4.0
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.21
ml_dtypes==0.5.1
mpmath==1.3.0
msgpack==1.1.0
namex==0.0.9
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.9.1
numpy==1.26.4
oauthlib==3.2.2
openai==1.79.0
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.15.0
packaging==25.0
pandas==2.2.3
parso==0.8.4
patsy==1.0.1
pillow==11.2.1
platformdirs==4.3.8
prompt_toolkit==3.0.51
protobuf==4.25.7
psutil==7.0.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycountry==24.6.1
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
pywin32==310
PyYAML==6.0.2
pyzmq==26.4.0
qbstyles==0.1.4
ray==2.46.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==14.0.0
rpds-py==0.25.1
rsa==4.9.1
safetensors==0.5.3
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
sentencepiece==0.2.0
shapely==2.0.7
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
sounddevice==0.5.2
stack-data==0.6.3
starlette==0.46.2
statsmodels==0.14.4
sumy==0.11.0
svgwrite==1.4.3
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.19.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_keras==0.1
termcolor==3.1.0
threadpoolctl==3.6.0
tifffile==2024.8.30
tokenizers==0.21.1
torch==2.7.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.52.3
Tree==0.2.4
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
wcwidth==0.2.13
Werkzeug==2.3.6
wrapt==1.17.2
xyzservices==2025.4.0
zipp==3.21.0
